Version 2.0.0 - Upgrade to Livewire 3 - And remove jetstream view files. Every new project should publish jetstream view files separately. - Re-arrange view files to minimise clash with project's jetstream files. - Use vite instead of laravel-mix - Conditionally add guest members to the project using config

+++ database/migrations/2023_02_05_111058_create_guests_table.php
@@ -8,15 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('guests', function (Blueprint $table) {
-            $table->id();
-            $table->ulid('token');
-            $table->timestamps();
-        });
+        if (config('quepenny.guest_members')) {
+            Schema::create('guests', function (Blueprint $table) {
+                $table->id();
+                $table->ulid('token');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::dropIfExists('guests');
+        if (config('quepenny.guest_members')) {
+            Schema::dropIfExists('guests');
+        }
     }
 };
